+++ Firestore/core/src/firebase/firestore/immutable/array_sorted_map.h
@@ -26,6 +26,7 @@
 
 #include "Firestore/core/src/firebase/firestore/immutable/map_entry.h"
 #include "Firestore/core/src/firebase/firestore/immutable/sorted_map_base.h"
+#include "Firestore/core/src/firebase/firestore/util/comparison.h"
 #include "Firestore/core/src/firebase/firestore/util/firebase_assert.h"
 
 namespace firebase {
@@ -118,7 +119,7 @@ class FixedArray {
  * ArraySortedMap is a value type containing a map. It is immutable, but has
  * methods to efficiently create new maps that are mutations of it.
  */
-template <typename K, typename V, typename C = std::less<K>>
+template <typename K, typename V, typename C = util::Comparator<K>>
 class ArraySortedMap : public SortedMapBase {
  public:
   using key_comparator_type = KeyComparator<K, V, C>;
@@ -227,6 +228,10 @@ class ArraySortedMap : public SortedMapBase {
     } else {
       return not_found;
     }
+  }
+
+  const key_comparator_type& comparator() const {
+    return key_comparator_;
   }
 
   /** Returns true if the map contains no elements. */