+++ Firestore/CHANGELOG.md
@@ -1,4 +1,8 @@
 # Unreleased
+- [feature] Added FieldValue.arrayUnion() and FieldValue.arrayRemove() methods
+  which can be used inside setData() or updateData() calls to atomically add
+  or remove specific elements to an array field in a document without using a
+  transaction.
 - [changed] Replaced the `DocumentListenOptions` object with a simple boolean.
   Instead of calling
   `addSnapshotListener(options: DocumentListenOptions.includeMetadataChanges(true))`
@@ -16,6 +20,10 @@
 - [changed] Replaced the `SetOptions` object with a simple boolean. Instead of
   calling `setData(["a": "b"], options: SetOptions.merge())` call
   `setData(["a": "b"], merge: true)`.
+- [changed] Replaced the `SnapshotOptions` object with direct use of the
+  `FIRServerTimestampBehavior` on `DocumentSnapshot`. Instead of calling
+  `data(SnapshotOptions.serverTimestampBehavior(.estimate))` call
+  `data(serverTimestampBehavior: .estimate)`. Changed `get` similarly.
 - [changed] Added ability to control whether DocumentReference.getDocument() and
   Query.getDocuments() should fetch from server only, cache only, or attempt
   server and fall back to the cache (which was the only option previously, and